make a directory for views
