+++ mercury/views/simulator.py
@@ -1,17 +1,8 @@
 from django.shortcuts import render
 from django.views.generic import TemplateView
 from django.http import HttpResponseRedirect
-from .forms import VehicleForm
+from ..forms import VehicleForm
 import datetime
-
-
-class HomePageView(TemplateView):
-    def get(self, request, **kwargs):
-        return render(request, "index.html", context=None)
-
-
-class AboutPageView(TemplateView):
-    template_name = "about.html"
 
 
 class SimulatorView(TemplateView):
@@ -28,7 +19,3 @@ class SimulatorView(TemplateView):
     def get(self, request, *args, **kwargs):
         form = self.form_class(initial={"created_at": datetime.datetime.now()})
         return render(request, "simulator.html", {"form": form})
-
-
-class DashboardView(TemplateView):
-    template_name = "dashboard.html"
